# 4.07.23(2)/4.07.23(2)/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "4.07.23(2)",
    "version": "1.0"
  },
  "paths": {
...
    "/Default": {
      "get": {
        "tags": [
          "Default"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Default"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSeeker"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSeeker"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/JobSeeker"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Default"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IsSahibi"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IsSahibi"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IsSahibi"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Default/{id}": {
      "get": {
        "tags": [
          "Default"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/FirmaAPI/{id}": {
      "get": {
        "tags": [
          "FirmaAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "FirmaAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Firma"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Firma"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Firma"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "FirmaAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FirmaAPI": {
      "post": {
        "tags": [
          "FirmaAPI"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirSek"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FirSek"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FirSek"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FirmaSektor": {
      "get": {
        "tags": [
          "FirmaSektor"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FirmaSektor"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FirmaSektor"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FirmaSektor"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FirmaSektor"
        ],
        "parameters": [
          {
            "name": "firmaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sektorId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "FirmaSektor"
        ],
        "parameters": [
          {
            "name": "firmaid",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sektorid",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sektor"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Sektor"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Sektor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FirmaSektor/{id}": {
      "get": {
        "tags": [
          "FirmaSektor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FirmaSektor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/IlansAPI/{id}": {
      "get": {
        "tags": [
          "IlansAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Ilan"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ilan"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ilan"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "IlansAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ilan"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Ilan"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Ilan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "IlansAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Sektor/{id}": {
      "get": {
        "tags": [
          "Sektor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sektor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Sektor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CalismaTuru": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ad": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Deneyim": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ad": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FirSek": {
        "type": "object",
        "properties": {
          "firma": {
            "$ref": "#/components/schemas/Firma"
          },
          "sektor": {
            "$ref": "#/components/schemas/Sektor"
          }
        },
        "additionalProperties": false
      },
      "Firma": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ad": {
            "type": "string",
            "nullable": true
          },
          "adres": {
            "type": "string",
            "nullable": true
          },
          "mail": {
            "type": "string",
            "nullable": true
          },
          "kurulusYili": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "telno": {
            "type": "integer",
            "format": "int64"
          },
          "aciklama": {
            "type": "string",
            "nullable": true
          },
          "isSahibiId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FirmaSektor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firmaId": {
            "type": "integer",
            "format": "int32"
          },
          "firma": {
            "$ref": "#/components/schemas/Firma"
          },
          "sektorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sektor": {
            "$ref": "#/components/schemas/Sektor"
          }
        },
        "additionalProperties": false
      },
      "Ilan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "aciklama": {
            "type": "string",
            "nullable": true
          },
          "deneyimId": {
            "type": "integer",
            "format": "int32"
          },
          "deneyim": {
            "$ref": "#/components/schemas/Deneyim"
          },
          "getDate": {
            "type": "string",
            "format": "date-time"
          },
          "egitimseviyesi": {
            "type": "string",
            "nullable": true
          },
          "maaş": {
            "type": "integer",
            "format": "int32"
          },
          "firmaId": {
            "type": "integer",
            "format": "int32"
          },
          "firma": {
            "$ref": "#/components/schemas/Firma"
          },
          "firmaWebSite": {
            "type": "string",
            "nullable": true
          },
          "calismaTuruId": {
            "type": "integer",
            "format": "int32"
          },
          "calismaTuru": {
            "$ref": "#/components/schemas/CalismaTuru"
          }
        },
        "additionalProperties": false
      },
      "IsSahibi": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ad": {
            "type": "string",
            "nullable": true
          },
          "soyisim": {
            "type": "string",
            "nullable": true
          },
          "mail": {
            "type": "string",
            "nullable": true
          },
          "telno": {
            "type": "integer",
            "format": "int64"
          },
          "kullanıcıSifresi": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "JobSeeker": {
        "required": [
          "kullanıcıSifresi"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ad": {
            "type": "string",
            "nullable": true
          },
          "soyad": {
            "type": "string",
            "nullable": true
          },
          "mail": {
            "type": "string",
            "nullable": true
          },
          "telno": {
            "type": "integer",
            "format": "int64"
          },
          "sehir": {
            "type": "string",
            "nullable": true
          },
          "ilce": {
            "type": "string",
            "nullable": true
          },
          "dogumtarihi": {
            "type": "string",
            "format": "date-time"
          },
          "egitimSeviyesi": {
            "type": "string",
            "nullable": true
          },
          "cinsiyet": {
            "type": "string",
            "nullable": true
          },
          "kullanıcıSifresi": {
            "minLength": 1,
            "type": "string"
          },
          "role": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Sektor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ad": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}